jjgt_venpermuta/requirements.txt
streamlit>=1.32.0
Pillow>=10.0.0
pandas>=2.0.0
gspread>=6.0.0
google-auth>=2.27.0
toml>=0.10.2
openpyxl>=3.1.0
